Receptor/Monitor/requirements.txt
paho-mqtt>=1.6.0
pyserial>=3.5
flask>=2.0.0
flask-socketio>=5.0.0
